--- seata-spring-boot-starter/src/main/java/io/seata/spring/boot/autoconfigure/HttpAutoConfiguration.java
@@ -1,51 +1,0 @@
-/*
- *  Copyright 1999-2019 Seata.io Group.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-package io.seata.spring.boot.autoconfigure;
-
-import io.seata.integration.http.HttpHandlerExceptionResolver;
-import io.seata.integration.http.TransactionPropagationInterceptor;
-import io.seata.integration.http.WebMvcConfigurerAdapter;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.HandlerExceptionResolver;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-
-import java.util.List;
-
-import static io.seata.spring.boot.autoconfigure.StarterConstants.HTTP_PREFIX;
-
-/**
- * Auto bean add for spring context if in springboot env.
- *
- * @author wangxb
- */
-@Configuration(proxyBeanMethods = false)
-@ConditionalOnWebApplication
-@ConditionalOnProperty(prefix = HTTP_PREFIX, name = "autoconfigure-enabled", havingValue = "true", matchIfMissing = true)
-public class HttpAutoConfiguration implements WebMvcConfigurerAdapter {
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(new TransactionPropagationInterceptor());
-    }
-
-    @Override
-    public void extendHandlerExceptionResolvers(List<HandlerExceptionResolver> exceptionResolvers) {
-        exceptionResolvers.add(new HttpHandlerExceptionResolver());
-    }
-
-}
