+++ seata-spring-boot-starter/src/main/java/io/seata/spring/boot/autoconfigure/SeataHttpAutoConfiguration.java
@@ -1,0 +1,38 @@
+/*
+ *  Copyright 1999-2019 Seata.io Group.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package io.seata.spring.boot.autoconfigure;
+
+import io.seata.integration.http.SeataWebMvcConfigurer;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+/**
+ * Auto bean add for spring context if in springboot env.
+ *
+ * @author wangxb
+ */
+@Configuration(proxyBeanMethods = false)
+@ConditionalOnWebApplication
+@ConditionalOnMissingBean(SeataWebMvcConfigurer.class)
+public class SeataHttpAutoConfiguration {
+
+    @Bean
+    public SeataWebMvcConfigurer seataWebMvcConfigurer() {
+        return new SeataWebMvcConfigurer();
+    }
+}
