--- seata-spring-boot-starter/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -1,12 +1,0 @@
-{
-  "groups": [
-  ],
-  "properties": [
-    {
-      "name": "seata.http.autoconfigure-enabled",
-      "type": "java.lang.Boolean",
-      "sourceType": "io.seata.spring.boot.autoconfigure.HttpAutoConfiguration",
-      "defaultValue": true
-    }
-  ]
-}
