+++ seata-spring-boot-starter/src/main/resources/META-INF/spring.factories
@@ -1,6 +1,6 @@
 # Auto Configure
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-io.seata.spring.boot.autoconfigure.SeataDataSourceAutoConfiguration,\
 io.seata.spring.boot.autoconfigure.SeataAutoConfiguration,\
-io.seata.spring.boot.autoconfigure.HttpAutoConfiguration,\
+io.seata.spring.boot.autoconfigure.SeataDataSourceAutoConfiguration,\
+io.seata.spring.boot.autoconfigure.SeataHttpAutoConfiguration,\
 io.seata.spring.boot.autoconfigure.SeataSagaAutoConfiguration
