+++ seata-spring-boot-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,4 +1,4 @@
-io.seata.spring.boot.autoconfigure.SeataDataSourceAutoConfiguration
 io.seata.spring.boot.autoconfigure.SeataAutoConfiguration
-io.seata.spring.boot.autoconfigure.HttpAutoConfiguration
+io.seata.spring.boot.autoconfigure.SeataDataSourceAutoConfiguration
+io.seata.spring.boot.autoconfigure.SeataHttpAutoConfiguration
 io.seata.spring.boot.autoconfigure.SeataSagaAutoConfiguration
